services/tsg-officer/pyproject.toml
[project]
name = "tsg-officer"
version = "0.1.0"
description = "LangGraph + Streamlit scaffold for a virtual TSG Officer approval copilot."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "langgraph>=1.0.0",
  "langgraph-checkpoint-sqlite>=3.0.0",
  "streamlit>=1.30.0",
  "requests>=2.0.0",
  "pydantic>=2.0.0",
  "jsonschema>=4.0.0",
  "pyyaml>=6.0.0",
  "python-dotenv>=1.0.0",
  "typing-extensions>=4.0.0",
]

[project.optional-dependencies]
openai = [
  "langchain-openai>=1.0.0",
  "langchain-core>=0.2.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["tsg_officer*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
